refactor: activity page
